requirements.txt
langchain
pymupdf
huggingface-hub
faiss-cpu
sentence-transformers
ipykernel
ipywidgets
arize-phoenix
arize-phoenix[experimental]
numpy
langchain
python-dev-tools
langchain-community
langchain-openai
tiktoken
